Add OptionSeed which implements DeserializeSeed for options
